scenario-3-keyvault/requirements.txt
flask
gunicorn
azure-keyvault-secrets
azure-identity
